% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomed_functions.R
\name{consensus_seq}
\alias{consensus_seq}
\title{Determine a consensus sequence}
\usage{
consensus_seq(seq, minpct = 0.5)
}
\arguments{
\item{seq}{XStringSet object}

\item{minpct}{real number}
}
\value{
vector with consensus letter for each position or "X" if no consensus reached
}
\description{
Determine a consensus sequence
}
\examples{
consensus_seq(Biostrings::XStringSet(c("ACGT", "ACGT", "ACGT", "ACGT")), .5)
}
