% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_functions.R
\name{get_r_package_info}
\alias{get_r_package_info}
\title{Get R package information}
\usage{
get_r_package_info()
}
\value{
A tibble with the following columns: package, version, and date.
}
\description{
This function returns a tibble with information about the R packages that are currently attached to the R session.
}
\examples{
get_r_package_info()

}
