% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jimsfunctions-package.R
\docType{package}
\name{jimsfunctions-package}
\alias{jimsfunctions}
\alias{jimsfunctions-package}
\title{jimsfunctions: Jim's Frequently Used Functions}
\description{
This is a small collection of the functions I frequently use for my research.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jameshunterbr/jimsfunctions}
  \item Report bugs at \url{https://github.com/jameshunterbr/jimsfunctions/issues}
}

}
\author{
\strong{Maintainer}: James Hunter \email{jameshunterbr@gmail.com}

}
\keyword{internal}
