% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_functions.R
\name{pval_aov}
\alias{pval_aov}
\title{Get p-value from an ANOVA model}
\usage{
pval_aov(model)
}
\arguments{
\item{model}{anova model}
}
\value{
p-value
}
\description{
Get p-value from an ANOVA model
}
\examples{
pval_aov(aov(mpg ~ cyl, mtcars))

}
