% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math_functions.R
\name{r2_aov}
\alias{r2_aov}
\title{Get R-squared from an ANOVA model}
\usage{
r2_aov(model)
}
\arguments{
\item{model}{anova model}
}
\value{
R-squared
}
\description{
Get R-squared from an ANOVA model
}
\examples{
r2_aov(aov(mpg ~ cyl, mtcars))

}
